feat: update layout and button styles, add PostCard component
