+++ components/Button.tsx
@@ -11,7 +11,7 @@ export default function Button(props: ButtonProps) {
   return (
     <div className={`relative ${className}`}>
       <button
-        className="absolute left-[10%] top-[15%] h-full w-full bg-primary-500 text-gray-500 transition-all hover:left-0 hover:top-0"
+        className="absolute left-[6%] top-[15%] h-full w-full bg-primary-500 text-gray-500 transition-all duration-300 hover:left-0 hover:top-0"
         onClick={onClick}
       >
         {children}
